requirements.txt
vllm
torch
numpy
hf_transfer
aiohttp
cupy-cuda12x
fastapi
uvicorn[standard]